transfer calculations is now float

+++ src/main/kotlin/repository/Repository.kt
@@ -26,7 +26,7 @@ data class Receipt(
     val amount: Float
 )
 
-data class Transfer(val from: User, val to: User, val amount: Int)
+data class Transfer(val from: User, val to: User, val amount: Float)
 
 interface ReceiptRepo {
     fun addReceipt(chatId: Long, from: UserId, to: List<UserId>, fullAmount: Float): Receipt
